requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
decorator==4.4.2
distro==1.9.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
lxml==5.0.0
moviepy==1.0.3
numpy==1.24.4
openai==1.6.1
pillow==10.2.0
proglog==0.1.10
pydantic==2.5.3
pydantic-core==2.14.6
pydub==0.25.1
python-docx==1.1.0
pytube==15.0.0
requests==2.31.0
sniffio==1.3.0
tqdm==4.66.1
typing-extensions==4.9.0
urllib3==2.1.0
